requirements/base.txt
django==2.1
djangorestframework==3.8.2
django-environ==0.4.5